+++ nix/overlay.nix
@@ -23,9 +23,10 @@ with final.haskell.lib;
     in
     {
       "sydtest" = sydtestPkg "sydtest";
-      "sydtest-hspec" = sydtestPkg "sydtest-hspec";
       "sydtest-aeson" = sydtestPkg "sydtest-aeson";
       "sydtest-discover" = sydtestPkgWithOwnComp "sydtest-discover";
+      "sydtest-hedgehog" = sydtestPkg "sydtest-hedgehog";
+      "sydtest-hspec" = sydtestPkg "sydtest-hspec";
       "sydtest-persistent" = sydtestPkg "sydtest-persistent";
       "sydtest-persistent-sqlite" = sydtestPkg "sydtest-persistent-sqlite";
       "sydtest-process" = sydtestPkg "sydtest-process";